+++ index.js
@@ -10,21 +10,6 @@ require('angular-material');
 
 require ('./dist/mdPickers');
 
-// Dayjs
-const dayjs = require('dayjs')
-var localeData = require('dayjs/plugin/localeData');
-var localizedFormat = require('dayjs/plugin/localizedFormat');
-var customParseFormat = require('dayjs/plugin/customParseFormat');
-var isSameOrAfter = require('dayjs/plugin/isSameOrAfter');
-var isSameOrBefore = require('dayjs/plugin/isSameOrBefore');
-var utc = require('dayjs/plugin/utc')
-var timezone = require('dayjs/plugin/timezone')
-dayjs.extend(localeData);
-dayjs.extend(localizedFormat);
-dayjs.extend(customParseFormat);
-dayjs.extend(isSameOrAfter);
-dayjs.extend(isSameOrBefore);
-dayjs.extend(utc)
-dayjs.extend(timezone)
+require ('moment-timezone');
 
 module.exports = 'mdPickers';