+++ src/components/mdpTimePicker/mdpTimePicker.js
@@ -1,23 +1,22 @@
-/* global dayjs, angular */
+/* global moment, angular */
 
 function TimePickerCtrl($scope, $mdDialog, time, autoSwitch, ampm, $mdMedia, options) {
     var self = this;
 
-    var settings = options.settings;
-
+    console.log(options)
 
     this.VIEW_HOURS = 1;
     this.VIEW_MINUTES = 2;
     this.currentView = this.VIEW_HOURS;
-    this.time = dayjs(time);
-    this.autoSwitch = !!autoSwitch;
+    this.time = extendedMoment({attrs: [time], disableTimezone: options.settings.disableTimezone});
+    this.time.setLocale(options.settings.currentLocale);
     this.ampm = !!ampm;
 
     this.hoursFormat = self.ampm ? "h" : "H";
     this.minutesFormat = "mm";
 
-    this.clockHours = parseInt(this.time.format(this.hoursFormat));
-    this.clockMinutes = parseInt(this.time.format(this.minutesFormat));
+    this.clockHours = parseInt(this.time.moment.format(this.hoursFormat));
+    this.clockMinutes = parseInt(this.time.moment.format(this.minutesFormat));
 
     $scope.$mdMedia = $mdMedia;
 
@@ -26,13 +25,13 @@ function TimePickerCtrl($scope, $mdDialog, time, autoSwitch, ampm, $mdMedia, opt
     };
 
     this.setAM = function() {
-        if(self.time.hour() >= 12)
-            self.time = self.time.hour(self.time.hour() - 12);
+        if(self.time.moment.hours() >= 12)
+            self.time.moment.hour(self.time.moment.hour() - 12);
     };
 
     this.setPM = function() {
-        if(self.time.hour() < 12)
-            self.time = self.time.hour(self.time.hour() + 12);
+        if(self.time.moment.hours() < 12)
+            self.time.moment.hour(self.time.moment.hour() + 12);
     };
 
     this.cancel = function() {
@@ -40,7 +39,7 @@ function TimePickerCtrl($scope, $mdDialog, time, autoSwitch, ampm, $mdMedia, opt
     };
 
     this.confirm = function() {
-        $mdDialog.hide(this.time);
+        $mdDialog.hide(this.time.moment.toDate());
     };
 }
 
@@ -69,7 +68,7 @@ function ClockCtrl($scope) {
                 for(var i = f; i <= t; i+=f)
                     self.steps.push(i);
                 if (!self.ampm) self.steps.push(0);
-                self.selected = self.time.hour() || 0;
+                self.selected = self.time.moment.hours() || 0;
                 if(self.ampm && self.selected > 12) self.selected -= 12;
 
                 break;
@@ -77,7 +76,7 @@ function ClockCtrl($scope) {
                 for(var i = 5; i <= 55; i+=5)
                     self.steps.push(i);
                 self.steps.push(0);
-                self.selected = self.time.minute() || 0;
+                self.selected = self.time.moment.minutes() || 0;
 
                 break;
         }
@@ -126,11 +125,11 @@ function ClockCtrl($scope) {
         switch(self.type) {
             case TYPE_HOURS:
                 if(self.ampm && self.time.format("A") === "PM") time += 12;
-                this.time = this.time.hour(time);
+                this.time.moment.hours(time);
                 break;
             case TYPE_MINUTES:
                 if(time > 59) time -= 60;
-                this.time = this.time.minute(time);
+                this.time.moment.minutes(time);
                 break;
         }
 
@@ -213,19 +212,12 @@ module.provider("$mdpTimePicker", function() {
 
     this.$get = ["$mdDialog", "$mdpLocale", function($mdDialog, $mdpLocale) {
         return function (time, options) {
-            var settings = options.settings;
-            if (!dayjs(time).isValid()) {
-                time = dayjs().locale(settings.currentLocale);
-                if (settings.overrideTimezone) {
-                    time = time.tz(settings.timezone, true)
-                } else {
-                    time = time.tz(dayjs.tz.guess(), true)
-                }
-            } else {
-                time = dayjs(time);
-            }
+            options.settings = options.settings || {disableTimezone: false, currentLocale: "en"};
 
             if (!angular.isObject(options)) options = {};
+            if (!angular.isDate(time)) {
+                time = newDate(options.settings);
+            }
 
             var labelOk = options.okLabel || $mdpLocale.time.okLabel || LABEL_OK;
             var labelCancel = options.cancelLabel || $mdpLocale.time.cancelLabel || LABEL_CANCEL;
@@ -238,14 +230,14 @@ module.provider("$mdpTimePicker", function() {
                     '<md-dialog-content layout-gt-xs="row" layout-wrap>' +
                     '<md-toolbar layout-gt-xs="column" layout-xs="row" layout-align="center center" flex class="mdp-timepicker-time md-hue-1 md-primary">' +
                     '<div class="mdp-timepicker-selected-time">' +
-                    '<span ng-class="{ \'active\': timepicker.currentView == timepicker.VIEW_HOURS }" ng-click="timepicker.currentView = timepicker.VIEW_HOURS">{{ timepicker.time.format(timepicker.hoursFormat) }}</span>:' +
+                    '<span ng-class="{ \'active\': timepicker.currentView == timepicker.VIEW_HOURS }" ng-click="timepicker.currentView = timepicker.VIEW_HOURS">{{ timepicker.time.moment.format(timepicker.hoursFormat) }}</span>:' +
 
-                    '<span ng-class="{ \'active\': timepicker.currentView == timepicker.VIEW_MINUTES }" ng-click="timepicker.currentView = timepicker.VIEW_MINUTES">{{ timepicker.time.format(timepicker.minutesFormat) }}</span>' +
+                    '<span ng-class="{ \'active\': timepicker.currentView == timepicker.VIEW_MINUTES }" ng-click="timepicker.currentView = timepicker.VIEW_MINUTES">{{ timepicker.time.moment.format(timepicker.minutesFormat) }}</span>' +
                     '</div>' +
                     '<div layout="column" ng-show="timepicker.ampm" class="mdp-timepicker-selected-ampm">' +
 
-                    '<span ng-click="timepicker.setAM()" ng-class="{ \'active\': timepicker.time.hour() < 12 }">AM</span>' +
-                    '<span ng-click="timepicker.setPM()" ng-class="{ \'active\': timepicker.time.hour() >= 12 }">PM</span>' +
+                    '<span ng-click="timepicker.setAM()" ng-class="{ \'active\': timepicker.time.moment.hours() < 12 }">AM</span>' +
+                    '<span ng-click="timepicker.setPM()" ng-class="{ \'active\': timepicker.time.moment.hours() >= 12 }">PM</span>' +
                     '</div>' +
 
                     '</md-toolbar>' +
@@ -267,8 +259,7 @@ module.provider("$mdpTimePicker", function() {
                 locals: {
                     time: time,
                     autoSwitch: options.autoSwitch,
-                    ampm: angular.isDefined(options.ampm) ? options.ampm : $mdpLocale.time.ampm,
-                    options: options
+                    ampm: angular.isDefined(options.ampm) ? options.ampm : $mdpLocale.time.ampm, options: options
                 },
                 multiple: true,
                 parent: PARENT_GETTER()
@@ -279,14 +270,14 @@ module.provider("$mdpTimePicker", function() {
 
 function compareTimeValidator(scope, value, format, otherTime, comparator) {
     // take only the date part, not the time part
-    if (dayjs(otherTime).isValid()) {
-        otherTime = dayjs(otherTime).format(format);
+    if (angular.isDate(otherTime)) {
+        otherTime = extendedMoment({attrs: [otherTime], disableTimezone: scope.settings.disableTimezone}).moment.format(format);
     }
-    otherTime = dayjs(otherTime, format, scope.settings.currentLocale, true);
-    var date = dayjs(value).isValid() ? dayjs(value) :  dayjs(value, format, scope.settings.currentLocale, true);
+    otherTime = extendedMoment({attrs: [otherTime, format, true], disableTimezone: scope.settings.disableTimezone}).moment;
+    var date = angular.isDate(value) ? extendedMoment({attrs: [value], disableTimezone: scope.settings.disableTimezone}).moment :  extendedMoment({attrs: [value, format, true], disableTimezone: scope.settings.disableTimezone}).moment;
 
     return !value ||
-            dayjs(value).isValid() ||
+            angular.isDate(value) ||
             !otherTime.isValid() ||
             comparator(date, otherTime);
 }
@@ -383,18 +374,11 @@ module.directive("mdpTimePicker", ["$mdpTimePicker", "$timeout", "$mdpLocale", f
 
             // update input element if model has changed
             ngModel.$formatters.unshift(function(value) {
-                var time = null;
-                if (dayjs(value).isValid()) {
-                    var time = dayjs(value);
-                    if (settings.overrideTimezone) {
-                        time = time.tz(settings.timezone, true);
-                    }
-                }
-
+                var time = angular.isDate(value) && extendedMoment({attrs: [value], disableTimezone: scope.settings.disableTimezone}).setLocale(scope.settings.currentLocale).moment;
                 if(time && time.isValid()) {
                     var strVal = time.format(scope.timeFormat);
                     updateInputElement(strVal);
-                    return strVal;
+                    return null;
                 } else {
                     updateInputElement(null);
                     return null;
@@ -402,7 +386,7 @@ module.directive("mdpTimePicker", ["$mdpTimePicker", "$timeout", "$mdpLocale", f
             });
 
             ngModel.$validators.format = function(modelValue, viewValue) {
-                return !viewValue || dayjs(viewValue).isValid() || dayjs(viewValue, scope.timeFormat, scope.settings.currentLocale, true).isValid();
+                return !viewValue || angular.isDate(viewValue) || extendedMoment({attrs: [viewValue, scope.timeFormat], disableTimezone: scope.settings.disableTimezone}).setLocale(scope.settings.currentLocale).moment.isValid();
             };
 
             ngModel.$validators.required = function(modelValue, viewValue) {
@@ -418,21 +402,20 @@ module.directive("mdpTimePicker", ["$mdpTimePicker", "$timeout", "$mdpLocale", f
             };
 
             ngModel.$parsers.unshift(function(value) {
-                var parsed = dayjs(value, scope.timeFormat, scope.settings.currentLocale, true);
+                var parsed = extendedMoment({attrs: [value], disableTimezone: scope.settings.disableTimezone}).setLocale(scope.settings.currentLocale).moment;
 
                 if(parsed.isValid()) {
-                    if(dayjs(ngModel.$modelValue).isValid()) {
-                        var originalModel = dayjs(ngModel.$modelValue);
-                        originalModel = originalModel.minute(parsed.minute());
-                        originalModel = originalModel.hour(parsed.hour());
-                        originalModel = originalModel.second(parsed.second());
-
+                    if(angular.isDate(ngModel.$modelValue)) {
+                        var originalModel = extendedMoment({attrs: [ngModel.$modelValue], disableTimezone: scope.settings.disableTimezone}).setLocale(scope.settings.currentLocale).moment;
+                        originalModel.minutes(parsed.minutes());
+                        originalModel.hours(parsed.hours());
+                        originalModel.seconds(parsed.seconds());
 
                         parsed = originalModel;
                     }
 
-                    // Returns Dayjs object to the model (toDate)
-                    return parsed;
+                    return parsed.toDate();
+
                 } else
                     return null;
             });
@@ -443,47 +426,44 @@ module.directive("mdpTimePicker", ["$mdpTimePicker", "$timeout", "$mdpLocale", f
                 inputContainerCtrl.setHasValue(!ngModel.$isEmpty(value));
             }
 
-            function updateTime(time, fromManualInput) {
-                var value = dayjs().locale(settings.currentLocale);
+            function updateTime(time, fromEvent) {
+                var value;
+                var ok = true;
+                if (fromEvent) {
+                    if (time.length === 4 && time.split(":").length === 1) {
+                        time = time.slice(0, 2) + ":" + time.slice(2, 4);
+                    }
+                    if (time === "") {
+                        var previousValue = ngModel.$modelValue;
+                        ok = false;
+                        if (previousValue === null) {
+                            previousValue = newDate(scope.settings)
+                            // No Time or Date was set before, so we set the current time
+                            ok = true;
+                        }
+                        value = extendedMoment({attrs: [previousValue], disableTimezone: scope.settings.disableTimezone}).setLocale(scope.settings.currentLocale).moment;
 
-                if (settings.overrideTimezone) {
-                    value = value.tz(settings.timezone, true);
-                } else {
-                    value = value.tz(dayjs.tz.guess(), true)
-                }
-
-                if (time !== null) {
-
-                    if (dayjs(time).isValid() && !fromManualInput) {
-                        value = dayjs(time);
                     } else {
-                        // Check if the locale is loaded
-                        if (angular.isDefined(dayjs.Ls[value.locale()])) {
-                            if (dayjs.Ls[value.locale()].formats.LT === "HH:mm") {
-                                if (time.length === 4 && time.split(":").length === 1) {
-                                    time = time.slice(0, 2) + ":" + time.slice(2, 4);
-                                }
-                            }
-                        }
-                        value = dayjs(time, scope.timeFormat, scope.settings.currentLocale);
-
-                        if (settings.overrideTimezone) {
-                            value = value.tz(scope.settings.timezone, true).local()
-                            value = value.tz(scope.settings.timezone)
-                        } else {
-                            value = value.tz(dayjs.tz.guess(), true)
-                        }
+                        value = extendedMoment({attrs: [], disableTimezone: scope.settings.disableTimezone}).setLocale(scope.settings.currentLocale).moment;
+                        value.minutes(time.split(":")[1]);
+                        value.hours(time.split(":")[0]);
+                    }
+                } else {
+                    if (angular.isDate(time)) {
+                        value = extendedMoment({attrs: [time, null, true], disableTimezone: scope.settings.disableTimezone}).setLocale(scope.settings.currentLocale).moment;
+                    } else {
+                        value = extendedMoment({attrs: [time, scope.timeFormat], disableTimezone: scope.settings.disableTimezone}).setLocale(scope.settings.currentLocale).moment
                     }
                 }
 
-
+                value.locale(scope.settings.currentLocale);
                 var strValue = value.format(scope.timeFormat);
 
-                if(value.isValid()) {
+                if(value.isValid() && ok) {
                     updateInputElement(strValue);
                     ngModel.$setViewValue(value);
                 } else {
-                    updateInputElement(ngModel.$viewValue);
+                    updateInputElement(strValue);
                 }
 
                 if(!ngModel.$pristine &&
@@ -503,8 +483,7 @@ module.directive("mdpTimePicker", ["$mdpTimePicker", "$timeout", "$mdpLocale", f
                     autoSwitch: scope.autoSwitch,
                     ampm: scope.ampm,
                     settings: opts.settings
-                }).then(function(time)
-                {
+                }).then(function(time){
                     updateTime(time);
                 }, function (error) {
                     if (opts.clearOnCancel) {
@@ -553,7 +532,7 @@ module.directive("mdpTimePicker", ["$mdpTimePicker", "$timeout", function($mdpTi
                     cancelLabel: scope.cancelLabel,
                     ampm: scope.ampm
                 }).then(function(time) {
-                    ngModel.$setViewValue(dayjs(time).format(scope.format));
+                    ngModel.$setViewValue(extendedMoment({attrs: [time], disableTimezone: scope.settings.disableTimezone}).setLocale(scope.settings.currentLocale).moment.format(scope.format));
                     ngModel.$render();
                 });
             };